a_vision/requirements.txt
ultralytics>=8.0.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
mysql-connector-python>=8.0.0
face-recognition>=1.3.0
dlib>=19.24.0 